+++ include/turtlecoin-crypto.h
@@ -56,9 +56,9 @@ namespace Core
             static std::string cn_turtle_lite_slow_hash_v1(const std::string data);
             static std::string cn_turtle_lite_slow_hash_v2(const std::string data);
 
-            static std::string cn_soft_shell_slow_hash_v0(const std::string data, const uint32_t height);
-            static std::string cn_soft_shell_slow_hash_v1(const std::string data, const uint32_t height);
-            static std::string cn_soft_shell_slow_hash_v2(const std::string data, const uint32_t height);
+            static std::string cn_soft_shell_slow_hash_v0(const std::string data, const uint64_t height);
+            static std::string cn_soft_shell_slow_hash_v1(const std::string data, const uint64_t height);
+            static std::string cn_soft_shell_slow_hash_v2(const std::string data, const uint64_t height);
 
             static std::string chukwa_slow_hash(const std::string data);
