+++ src/slow-hash-x86.c
@@ -421,7 +421,7 @@ BOOL SetLockPagesPrivilege(HANDLE hProcess, BOOL bEnable)
  * the allocated buffer.
  */
 
-void slow_hash_allocate_state(uint32_t page_size)
+void slow_hash_allocate_state(uint64_t page_size)
 {
     if(hp_state != NULL)
     {
@@ -457,7 +457,7 @@ void slow_hash_allocate_state(uint32_t page_size)
  *@brief frees the state allocated by slow_hash_allocate_state
  */
 
-void slow_hash_free_state(uint32_t page_size)
+void slow_hash_free_state(uint64_t page_size)
 {
     if(hp_state == NULL)
     {
@@ -511,11 +511,11 @@ void slow_hash_free_state(uint32_t page_size)
  * @param length the length in bytes of the data
  * @param hash a pointer to a buffer in which the final 256 bit hash will be stored
  */
-void cn_slow_hash(const void *data, size_t length, char *hash, int light, int variant, int prehashed, uint32_t page_size, uint32_t scratchpad, uint32_t iterations)
+void cn_slow_hash(const void *data, size_t length, char *hash, int light, int variant, int prehashed, uint64_t page_size, uint64_t scratchpad, uint64_t iterations)
 {
-    uint32_t TOTALBLOCKS = (page_size / AES_BLOCK_SIZE);
-    uint32_t init_rounds = (scratchpad / INIT_SIZE_BYTE);
-    uint32_t aes_rounds = (iterations / 2);
+    uint64_t TOTALBLOCKS = (page_size / AES_BLOCK_SIZE);
+    uint64_t init_rounds = (scratchpad / INIT_SIZE_BYTE);
+    uint64_t aes_rounds = (iterations / 2);
     size_t lightFlag = (light ? 2: 1);
 
     RDATA_ALIGN16 uint8_t expandedKey[240];  /* These buffers are aligned to use later with SSE functions */
